refactor: copyTextToClipboard を lib/clipboard.ts に移動しフォールバックを改善
- features/posts/lib/ から lib/clipboard.ts に移動してレイヤリング違反を解消
- Clipboard API → execCommand の順でフォールバックし、デスクトップHTTP環境にも対応
- エラーメッセージを汎用的な "Failed to copy to clipboard" に変更
- ShareButton の handleCopyLink を copyTextToClipboard に統一
- 旧ファイルは再エクスポートで後方互換を維持
